+++ Source/Objects/Data Processing/Histogram Objects/PlotFFT/ORPlotFFT.m
@@ -47,7 +47,7 @@ NSString* ORPlotFFTShowChanged = @"ORPlotFFTShowChanged";
 
 
 #pragma mark ¥¥¥Accessors
-- (unsigned) numberChans
+- (NSUInteger) numberChans
 {
 	return [realArray count];
 }
